fixed the struct thing
so the compiler doesn't like talking about typedef or struct when they are not defined in the first tab.

I imgine we dont want a ton of struct defintions cluttering the main file so I created a header called struct.h that is called in the main file.

This push actually compiles.